Merge-in from VC8 sandbox - Pixelbuffers for DX7 (DynTex runs!) - getBuffer(face,mip) and getPixelBox(face,mip) now take size_t instead of int, as negative values make no sense - Add OgreHardwarePixelBuffer::blit convience function to blit entire contents of source PixelBuffer - PixelUtil::getBitMasks to query masks on native endian formats

+++ trunk/OgreMain/include/OgreImage.h
@@ -240,7 +240,7 @@ namespace Ogre {
         /**
          * Get a PixelBox encapsulating the image data of a mipmap
          */
-        PixelBox getPixelBox(int face = 0, int mipmap = 0) const;
+        PixelBox getPixelBox(size_t face = 0, size_t mipmap = 0) const;
 
 		enum Filter
 		{
